+++ src/backend/datasources/filters/BinaryFilter.cpp
@@ -441,7 +441,7 @@ void BinaryFilterPrivate::readDataFromDevice(QIODevice& device, AbstractDataSour
 		DEBUG("reading row " << i);
 		//prepend the index if required
 		if (createIndexEnabled)
-			static_cast<QVector<int>*>(dataContainer[0])->operator[](i) = i+1;
+			static_cast<QVector<qint64>*>(dataContainer[0])->operator[](i) = i+1;
 
 		for (int n = startColumn; n < m_actualCols; ++n) {
 			DEBUG("reading column " << n);
